add server utils
